moved LS specific stuff to languages-server folder
