+++ src/sprotty/languageserver/index.ts
@@ -17,6 +17,9 @@
 export * from './code-action-palette';
 export * from './completion-label-edit';
 export * from './delete';
+export * from './ls-theia-sprotty-connector';
+export * from './ls-theia-diagram-server';
+export * from './delete';
 export * from './workspace-edit-command';
 export * from './rename-label-edit';
 export * from './traceable';
