+++ src/sprotty/languageserver/rename-label-edit.ts
@@ -18,7 +18,7 @@ import { RenameRequest } from '@theia/languages/lib/browser';
 import { SingleTextInputDialog } from '@theia/core/lib/browser';
 import { inject, injectable } from "inversify";
 import { SLabel } from "sprotty/lib";
-import { LSTheiaDiagramServerProvider } from "../theia-diagram-server";
+import { LSTheiaDiagramServerProvider } from "./ls-theia-diagram-server";
 import { Traceable, getRange } from './traceable';
 
 @injectable()
