+++ src/sprotty/languageserver/workspace-edit-command.ts
@@ -17,7 +17,7 @@
 import { Workspace, WorkspaceEdit } from "@theia/languages/lib/browser";
 import { inject, injectable } from "inversify";
 import { Action, Command, CommandExecutionContext, CommandResult, TYPES } from "sprotty/lib";
-import { LSTheiaDiagramServer } from "../theia-diagram-server";
+import { LSTheiaDiagramServer } from "./ls-theia-diagram-server";
 
 @injectable()
 export abstract class AbstractWorkspaceEditCommand extends Command {
