+++ src/sprotty/theia-sprotty-connector.ts
@@ -15,11 +15,10 @@
  ********************************************************************************/
 
 import { QuickPickService, WidgetManager } from '@theia/core/lib/browser';
-import { ILanguageClient, Workspace } from '@theia/languages/lib/browser';
+import { Workspace } from '@theia/languages/lib/browser';
 import { ActionMessage, ExportSvgAction, ServerStatusAction } from 'sprotty/lib';
 import { DiagramManager } from '../theia/diagram-manager';
-import { DiagramWidget } from '../theia/diagram-widget';
-import { ActionMessageReceiver, DiagramLanguageClient } from '../theia/languageserver';
+import { DiagramLanguageClient } from '../theia/languageserver';
 import { TheiaDiagramServer } from './theia-diagram-server';
 import { TheiaFileSaver } from './theia-file-saver';
 import { EditorManager } from '@theia/editor/lib/browser';
@@ -50,66 +49,3 @@ export interface TheiaSprottyConnector {
     onMessageReceived(message: ActionMessage): void
 }
 
-/**
- * Connects sprotty DiagramServers to a Theia LanguageClientContribution.
- *
- * Used to tunnel sprotty actions to and from the sprotty server through
- * the LSP.
- *
- * Instances bridge the gap between the sprotty DI containers (one per
- * diagram) and a specific language client from the Theia DI container
- * (one per application).
- */
-export class LSTheiaSprottyConnector implements TheiaSprottyConnector, TheiaSprottyConnectorServices, ActionMessageReceiver {
-    private servers: TheiaDiagramServer[] = []
-
-    readonly diagramLanguageClient: DiagramLanguageClient
-    readonly fileSaver: TheiaFileSaver
-    readonly editorManager: EditorManager
-    readonly widgetManager: WidgetManager
-    readonly diagramManager: DiagramManager
-    readonly workspace?: Workspace
-    readonly quickPickService?: QuickPickService
-
-    constructor(services: TheiaSprottyConnectorServices) {
-        Object.assign(this, services)
-        this.diagramLanguageClient.connect(this)
-    }
-
-    connect(diagramServer: TheiaDiagramServer) {
-        this.servers.push(diagramServer)
-        diagramServer.connect(this)
-    }
-
-    disconnect(diagramServer: TheiaDiagramServer) {
-        const index = this.servers.indexOf(diagramServer)
-        if (index >= 0)
-            this.servers.splice(index, 0)
-        diagramServer.disconnect()
-        this.diagramLanguageClient.didClose(diagramServer.clientId)
-    }
-
-    save(uri: string, action: ExportSvgAction) {
-        this.fileSaver.save(uri, action)
-    }
-
-    showStatus(widgetId: string, status: ServerStatusAction): void {
-        const widget = this.widgetManager.getWidgets(this.diagramManager.id).find(w => w.id === widgetId)
-        if (widget instanceof DiagramWidget)
-            widget.setStatus(status)
-    }
-
-    sendMessage(message: ActionMessage) {
-        this.diagramLanguageClient.sendThroughLsp(message)
-    }
-
-    getLanguageClient(): Promise<ILanguageClient> {
-        return this.diagramLanguageClient.languageClient
-    }
-
-    onMessageReceived(message: ActionMessage): void {
-        this.servers.forEach(element => {
-            element.messageReceived(message)
-        })
-    }
-}
